Add support for "shuffling" furballs to the CLI.
It's sort of like the concept of Undertale corruptions -- that is,
shuffling various assets or parts of assets to get a game that's
playable but mostly nonsense. I figured it'd be fun to have, and adding
it to the CLI means it can be trivially used for any furball (not just
the core game).